// src/lib/types/filters/user.ts
export interface FilterUser {
    id?: number;
    email?: string;
    username?: string;
    password?: string;
    name?: string;
    phone?: string;
    province?: string;
    city?: string;
    district?: string;
    type?: number;
    status?: number;
    created_at?: string;
    updated_at?: string;
}